--- src/data/about/en.js
@@ -1,17 +1,0 @@
-const about = {
-    title: "Hello, I'm Ensar Makas.",
-    jobTitle:"Software Engineer",
-    content: `
-    I am passionate about the internet and internet technologies, continuously developing my skills in this field.
-    Throughout my university education, I took part in various activities and actively sought professional experience.
-    My goal is to develop performance-oriented solutions in the field of internet technologies, considering the industry's needs and ensuring top-notch user and developer experiences.
-    `,
-    sections: [
-        {
-            title: "",
-            description: ``,
-        }
-    ]
-};
-
-export {about}
