--- src/data/about/tr.js
@@ -1,13 +1,0 @@
-const about = {
-    title: "Merhaba, ben Ensar Makas.",
-    jobTitle:"Yazılım Mühendisi",
-    content: `
-    İnternet ve internet teknolojilerinde ilgili ve bu konuda kendimi geliştiren biriyim.
-    Üniversitemdeki eğitim sürecim boyunca çeşitli etkinliklerde görev aldım ve profesyonel anlamda tecrübe edinmeye gayret gösterdim.
-    Hedefim internet teknolojileri alanında sektörün ihtiyaçlarını gözeterek performans odaklı, kullanıcı ve geliştirici deneyimlerini üst düzeyde tutacak çözümler üretmek
-                    
-    `,
-    sections: []
-}
-
-export {about}
