--- src/data/educations/en.js
@@ -1,25 +1,0 @@
-const educations = [
-    {
-        university: "Ecole 42 Turkey",
-        branch: "",
-        startDate: "January 2022",
-        endDate: "Ongoing",
-        description: "Ecole 42 is an international software school. We develop projects by taking on different tasks at all levels based on peer learning."
-    },
-    {
-        university: "Beykent University",
-        branch: "Computer Engineering",
-        startDate: "August 2019",
-        endDate: "Ongoing",
-        description: "All my courses are completed, only my internship is pending."
-    },
-    {
-        university: "Captain Ahmet Erdogan Anatolian Imam Hatip High School",
-        branch: "",
-        startDate: "October 2014",
-        endDate: "June 2019",
-        description: ""
-    }
-];
-
-export {educations}
