--- src/data/educations/tr.js
@@ -1,25 +1,0 @@
-const educations = [
-    {
-        university: "Ecole 42 Türkiye",
-        branch: "",
-        startDate: "Ocak 2022",
-        endDate: "Devam",
-        description: "Ecole 42, Uluslarası yazılım okuludur. Akran öğrenmesi metodu esas alınarak her seviyede farklı görevler üstlenerek projeler geliştiriyoruz."
-    },
-    {
-        university: "Beykent Üniversitesi",
-        branch: "Bilgisayar Mühendisliği",
-        startDate: "Ağustos 2019",
-        endDate: "Devam",
-        description: "Tüm derslerim bitti, sadece stajım kaldı."
-    },
-    {
-        university: "Kaptan Ahmet Erdoğan Anadolu İmam Hatip Lisesi",
-        branch: "",
-        startDate: "Ekim 2014",
-        endDate: "Haziran 2019",
-        description: ""
-    },
-]
-
-export {educations}
