--- src/data/jobs/en.js
@@ -1,28 +1,0 @@
-const jobs = [
-    {
-        jobTitle: "Intern",
-        company: "Koç Sistem",
-        description: "During my two months here, I had the opportunity to work with a professional team. I managed the project following the Agile-Scrum principles while developing a shopping API application.",
-        startDate: "November 2022",
-        endDate: "January 2023",
-        learnedTechs: ["PHP", "Javascript", "HTML", "CSS", "Wordpress", ".NET", ".NET Core"]
-    },
-    {
-        jobTitle: "Intern",
-        company: "Turkish Open Source Platform",
-        description: "During one month here, I had the opportunity to examine the software named LabelStudio.",
-        startDate: "July 2022",
-        endDate: "August 2022",
-        learnedTechs: ["LabelStudio", "DJango", "Python"]
-    },
-    {
-        jobTitle: "Volunteer Intern",
-        company: "Kando Informatics",
-        description: "Here, I gained my initial experiences in web applications. I learned languages such as PHP, MySQL, Javascript, HTML, CSS.",
-        startDate: "August 2020",
-        endDate: "May 2022",
-        learnedTechs: [".NET Core"]
-    }
-];
-
-export {jobs}
