--- src/data/jobs/tr.js
@@ -1,28 +1,0 @@
-const jobs = [
-    {
-        jobTitle: "Stajyer",
-        company: "Koç Sistem",
-        description:"Burada iki ay boyunca profesyonel bir ekiple çalışma fırsatı edindim. Bu süreçte alışveriş api uygulaması geliştirirken proje yönetimimi Agile-Scrum prensibine göre yaptım",
-        startDate:"Kasım-2022",
-        endDate:"Ocak-2023",
-        learnedTechs: ["PHP","Javascript","HTML","CSS","Wordpress",".NET", ".NET Core"]
-    },
-    {
-        jobTitle: "Stajyer",
-        company: "Türkiye Açık Kaynak Platformu",
-        description:"Burada bir ay boyunca LabelStudio isimli yazılımı inceleme fırsatı elde ettik",
-        startDate:"Temmuz-2022",
-        endDate:"Ağustos-2022",
-        learnedTechs: ["LabelStudio","DJango","Python"]
-    },
-    {
-        jobTitle: "Gönüllü Staj",
-        company: "Kando Bilişim",
-        description:"Burada internet uygulamaları alanında ilk tecrübelerimi edindim. PHP, MySQL, Javascript, HTML, CSS gibi dilleri öğrendim",
-        startDate:"Ağustos-2020",
-        endDate:"Mayıs-2022",
-        learnedTechs: [".NET Core"]
-    },
-]
-
-export {jobs}
