--- src/data/links/index.js
@@ -1,11 +1,0 @@
-const links = {
-    linkedIn: "https://www.linkedin.com/in/ensar-makas-a16b69199/",
-    github:"https://github.com/rasnesakam",
-    twitter:"https://twitter.com/ensar_makas",
-    gmail:"ensar.makas@gmail.com",
-    instagram:"https://www.instagram.com/ensar.makas/",
-    medium: "https://medium.com/@emakas",
-    others:[]
-};
-
-export {links}
