--- src/data/projects/en.js
@@ -1,52 +1,0 @@
-const projects = [
-    {
-        name: "DepIn",
-        description: "DepIn is a dependency injection tool designed for the Java language.",
-        github_link: "https://github.com/rasnesakam/DepIn",
-        techStack: ["Java", "Gradle", "JUnit"]
-    },
-    {
-        name: "Shop API",
-        description: "A product management service that will run background systems for e-commerce sites.",
-        github_link: "https://github.com/rasnesakam/shop-api",
-        techStack: ["C#", ".NET7", "Entity Framework", "Fluent API", "Docker"]
-    },
-    {
-        name: "emakas-users",
-        description: "A service created for user and authorization processes that I will process in my microservices. It contains user information and authorization processes.",
-        github_link: "https://github.com/rasnesakam/emakas-users",
-        techStack: ["Java", "Spring Boot", "Docker"]
-    },
-    {
-        name: "RESTful API Gateway",
-        description: "A gateway that will provide access to my microservices from a single location.",
-        github_link: "https://github.com/rasnesakam/api-gateway",
-        techStack: ["Typescript", "Node Js", "Express Js"]
-    },
-    {
-        name: "Blog API",
-        description: "Designed to manage backend processes for blog and news sites.",
-        github_link: "https://github.com/rasnesakam/blog-api",
-        techStack: ["Typescript", "Node Js", "Express Js", "Prisma Js"]
-    },
-    {
-        name: "Simple Dashboard",
-        description: "A common administration panel that can be used for my services. Users will access and add new content to their services from here.",
-        github_link: "https://github.com/rasnesakam/simple-dashboard",
-        techStack: ["Next Js"]
-    },
-    {
-        name: "Battery World",
-        description: "An example e-commerce application developed. Also, it is being developed for my friend.",
-        github_link: "https://github.com/rasnesakam/bataryadunyasi",
-        techStack: ["Next Js", "Docker"]
-    },
-    {
-        name: "Company Management System",
-        description: "A system based on MVC that focuses on task management for customer companies during my first job experiences.",
-        github_link: "https://github.com/rasnesakam/CompanyManagementSystem",
-        techStack: ["C#", ".NET5"]
-    }
-];
-
-export {projects}
