--- src/data/projects/tr.js
@@ -1,52 +1,0 @@
-const projects = [
-    {
-        name:"DepIn",
-        description:"DepIn Java dili için tasarlanan bir bağımlılık enjeksiyon aracıdır.",
-        github_link:"https://github.com/rasnesakam/DepIn",
-        techStack: ["Java","Gradle","JUnit"]
-    },
-    {
-        name:"Shop API",
-        description:"Alışveriş siteleri için arka plan sistemlerini yürütecek olan ürün yönetim servisidir.",
-        github_link:"https://github.com/rasnesakam/shop-api",
-        techStack: ["C#",".NET7","Entity Framework","Fluent API","Docker"]
-    },
-    {
-        name:"emakas-users",
-        description:"Mikro servislerimde işleyeceğim kullanıcı ve yetkilendirme işlemleri için oluşturduğum bir servis. Burada kullanıcı bilgileri ve yetkilendirme işlemleri yer alır.",
-        github_link:"https://github.com/rasnesakam/emakas-users",
-        techStack: ["Java","Spring Boot","Docker"]
-    },
-    {
-        name:"RESTful API Gateway",
-        description:"Mikroservislerim için tek bir yerden erişim mekanizması sağlayacak bir geçit.",
-        github_link:"https://github.com/rasnesakam/api-gateway",
-        techStack: ["Typescript","Node Js","Express Js"]
-    },
-    {
-        name:"Blog API",
-        description:"Blog ve haber siteleri için arka plan işlemlerini yönetecek bir servis olarak tasarlandı.",
-        github_link:"https://github.com/rasnesakam/blog-api",
-        techStack: ["Typescript","Node Js","Express Js","Prisma Js"]
-    },
-    {
-        name:"Simple Dashboard",
-        description:"Yazdığım servisler için ortak kullanılabilecek bir yönetim paneli. Kullanıcılar sahip oldukları servislerdeki içeriklere buradan erişip yeni içerikler ekleyecekler.",
-        github_link:"https://github.com/rasnesakam/simple-dashboard",
-        techStack: ["Next Js"]
-    },
-    {
-        name:"Batarya Dünyası",
-        description:"Geliştirilen örnek bir E ticaret uygulaması. Ayrıca arkadaşım için geliştirilmeye devam ediliyor",
-        github_link:"https://github.com/rasnesakam/bataryadunyasi",
-        techStack: ["Next Js","Docker"]
-    },
-    {
-        name:"Company Management System",
-        description:"İlk iş tecrübelerim sırasında oluşturulan, Müşteri şirketler için görev yönetimini hedef alan MVC tabanlı bir sistem",
-        github_link:"https://github.com/rasnesakam/CompanyManagementSystem",
-        techStack: ["C#",".NET5"]
-    },
-];
-
-export {projects}
