--- src/data/volunteerWorks/en.js
@@ -1,16 +1,0 @@
-const volunteerWorks = [
-    {
-        title: "Chairman of the Administrative Board, IEEE Beykent University Computer Society",
-        description: "During my term as the chairman of the administrative board, we organized competitions and seminars.",
-        startDate: "August 2021",
-        endDate: "July 2022"
-    },
-    {
-        title: "Active Member, IEEE Beykent University Student Branch",
-        description: "During my active membership, I participated in various committees and projects.",
-        startDate: "August 2020",
-        endDate: "July 2021"
-    }
-];
-
-export {volunteerWorks}
