--- src/data/volunteerWorks/tr.js
@@ -1,16 +1,0 @@
-const volunteerWorks = [
-    {
-        title: "IEEE Beykent Üniversitesi Computer Society İdari Kurul Başkanı",
-        description:"Pandemi döneminde gerçekleşen idari kurul başkanlığım boyunca yarışma ve seminerler düzenledik",
-        startDate:"Ağustos 2021",
-        endDate:"Temmuz 2022"
-    },
-    {
-        title: "IEEE Beykent Üniversitesi Öğrenci Kolu Aktif Üye",
-        description:"Aktif üyeliğim boyunca çeşitli komitelerdeki projelerde yer aldım",
-        startDate:"Ağustos 2020",
-        endDate:"Temmuz 2021"
-    }
-]
-
-export {volunteerWorks}
